src/processes/quality_control_processor.py
"""
Quality Control Processor - Example of a process with complex data validation and transformation.
This process demonstrates advanced data quality checks and cleaning operations.
"""
import pandas as pd
import numpy as np
from typing import Dict, Any, List, Tuple
from datetime import datetime
import re


class QualityControlProcessor:
    category = "quality_control"

    def process_data(self, inputs):
        """Sample quality control process. Returns missing value counts."""
        df = inputs.get('data')
        if df is None:
            return {"error": "No data provided"}
        missing = df.isnull().sum().to_dict()
        return {"missing_values": missing}

  